upload-service/requirements.txt
fastapi
uvicorn[standard]
slowapi
python-multipart
python-dotenv
boto3
celery[redis]
